requirements.txt
google-adk>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=0.19.0